Error check for message not modified.
In Telegram, if an update occurs but the message is the same, it
returns a 400 error. By the logic of the code this shouldn't happen
and I haven't been able to replicate it manually, but I've seen the
error occur so there must be some race condition or the like with
multiple users. This check just ignores it and prevents bot from
crashing.